+++ source/Erwin/debug/instrumentor.cpp
@@ -94,7 +94,7 @@ void InstrumentationTimer::stop()
     long long start = std::chrono::time_point_cast<std::chrono::microseconds>(start_timepoint_).time_since_epoch().count();
     long long end = std::chrono::time_point_cast<std::chrono::microseconds>(end_timepoint).time_since_epoch().count();
 
-    uint32_t thread_id = std::hash<std::thread::id>{}(std::this_thread::get_id());
+    size_t thread_id = std::hash<std::thread::id>{}(std::this_thread::get_id());
     Instrumentor::write_profile({ name_, start, end, thread_id });
 
     stopped_ = true;
